% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Hodge2019}
\alias{data_Hodge2019}
\title{Count matrix for single-cell transcriptomes for 3357
of the nuclei characterized in Hodge, et al. (2019). These data
are subset to 50 cells per donor to fit in GitHub. The complete
set of data is available at brain-map.org.  Note that this data
set has been re-aligned to a more recent transcriptome than what
was used in Hodge et al 2019.}
\format{
A matrix with 1809 columns (cells) and 24057 rows (genes)
\describe{
 \item{rows}{Each row corresponds to a single RefSeq gene symbol.}
 \item{columns}{Each column is named using a unique cell identifier that corresponds to the sample_name column in the metadata_Hodge2019 table.}
}
}
\usage{
data_Hodge2019
}
\description{
A matrix consisting of read count values for all 1809 cells
characterized in Hodge, et al. (2019).
}
\details{
Each row corresponds to a single gene symbol. Each column corresponds to a single cell.
}
\keyword{datasets}
