% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dend_Hodge2019}
\alias{dend_Hodge2019}
\title{Dendrogram of cluster organization}
\format{
An R dendrogram variable with 75 clusters and key and additional node and edge metadata
\describe{
 \item{dend_Hodge2019}{Dendorgram used in Hodge et al (2019).}
}
}
\usage{
dend_Hodge2019
}
\description{
An R dendrogram format for all 75 clusters
characterized in Hodge, et al. (2019).
}
\keyword{datasets}
